requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
hyperliquid-python-sdk==1.0.34
eth-account==0.10.0
python-dotenv==1.0.0
requests==2.31.0
websockets==15.0.1